refactor(ff-render): split nodes/composite.rs into per-node-type submodules
